requirements.txt
docker>=7.0.0
pyyaml>=6.0.1
rich>=13.7.0
pytest>=8.0.0
